+++ nvim/lua/custom/plugins/editor.lua
@@ -331,7 +331,9 @@ return {
       {
         '<C-a>',
         function()
-          if not vim.bo.modifiable then return end
+          if not vim.bo.modifiable then
+            return
+          end
           require('dial.map').manipulate('increment', 'normal')
         end,
         desc = 'Increment',
@@ -339,7 +341,9 @@ return {
       {
         '<C-x>',
         function()
-          if not vim.bo.modifiable then return end
+          if not vim.bo.modifiable then
+            return
+          end
           require('dial.map').manipulate('decrement', 'normal')
         end,
         desc = 'Decrement',
@@ -347,7 +351,9 @@ return {
       {
         'g<C-a>',
         function()
-          if not vim.bo.modifiable then return end
+          if not vim.bo.modifiable then
+            return
+          end
           require('dial.map').manipulate('increment', 'gnormal')
         end,
         desc = 'Increment (sequential)',
@@ -355,7 +361,9 @@ return {
       {
         'g<C-x>',
         function()
-          if not vim.bo.modifiable then return end
+          if not vim.bo.modifiable then
+            return
+          end
           require('dial.map').manipulate('decrement', 'gnormal')
         end,
         desc = 'Decrement (sequential)',
@@ -364,7 +372,9 @@ return {
         '<C-a>',
         mode = 'v',
         function()
-          if not vim.bo.modifiable then return end
+          if not vim.bo.modifiable then
+            return
+          end
           require('dial.map').manipulate('increment', 'visual')
         end,
         desc = 'Increment',
@@ -373,7 +383,9 @@ return {
         '<C-x>',
         mode = 'v',
         function()
-          if not vim.bo.modifiable then return end
+          if not vim.bo.modifiable then
+            return
+          end
           require('dial.map').manipulate('decrement', 'visual')
         end,
         desc = 'Decrement',
@@ -382,7 +394,9 @@ return {
         'g<C-a>',
         mode = 'v',
         function()
-          if not vim.bo.modifiable then return end
+          if not vim.bo.modifiable then
+            return
+          end
           require('dial.map').manipulate('increment', 'gvisual')
         end,
         desc = 'Increment (sequential)',
@@ -391,7 +405,9 @@ return {
         'g<C-x>',
         mode = 'v',
         function()
-          if not vim.bo.modifiable then return end
+          if not vim.bo.modifiable then
+            return
+          end
           require('dial.map').manipulate('decrement', 'gvisual')
         end,
         desc = 'Decrement (sequential)',